move geolocation to services
